@@ -163,4 +163,4 @@ require (
 
 toolchain go1.21.0
 
-go 1.20
+go 1.21
