@@ -2,8 +2,8 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-//go:build go1.13 && !windows
-// +build go1.13,!windows
+//go:build !windows
+// +build !windows
 
 package template
 
@@ -11,10 +11,11 @@ package template
 
 import (
 	"fmt"
-	"github.com/gohugoio/hugo/tpl/internal/go_templates/texttemplate/parse"
 	"os"
 	"strings"
 	"testing"
+
+	"github.com/gohugoio/hugo/tpl/internal/go_templates/texttemplate/parse"
 )
 
 const (
@@ -31,22 +32,32 @@ type multiParseTest struct {
 }
 
 var multiParseTests = []multiParseTest{
-	{"empty", "", noError,
+	{
+		"empty", "", noError,
 		nil,
-		nil},
-	{"one", `{{define "foo"}} FOO {{end}}`, noError,
+		nil,
+	},
+	{
+		"one", `{{define "foo"}} FOO {{end}}`, noError,
 		[]string{"foo"},
-		[]string{" FOO "}},
-	{"two", `{{define "foo"}} FOO {{end}}{{define "bar"}} BAR {{end}}`, noError,
+		[]string{" FOO "},
+	},
+	{
+		"two", `{{define "foo"}} FOO {{end}}{{define "bar"}} BAR {{end}}`, noError,
 		[]string{"foo", "bar"},
-		[]string{" FOO ", " BAR "}},
+		[]string{" FOO ", " BAR "},
+	},
 	// errors
-	{"missing end", `{{define "foo"}} FOO `, hasError,
+	{
+		"missing end", `{{define "foo"}} FOO `, hasError,
 		nil,
-		nil},
-	{"malformed name", `{{define "foo}} FOO `, hasError,
 		nil,
-		nil},
+	},
+	{
+		"malformed name", `{{define "foo}} FOO `, hasError,
+		nil,
+		nil,
+	},
 }
 
 func TestMultiParse(t *testing.T) {
@@ -432,7 +443,7 @@ func TestIssue19294(t *testing.T) {
 	// by the contents of "stylesheet", but if the internal map associating
 	// names with templates is built in the wrong order, the empty block
 	// looks non-empty and this doesn't happen.
-	var inlined = map[string]string{
+	inlined := map[string]string{
 		"stylesheet": `{{define "stylesheet"}}stylesheet{{end}}`,
 		"xhtml":      `{{block "stylesheet" .}}{{end}}`,
 	}
